+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ahincho/nova-nestjs
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @ahincho/nova-nestjs-observability@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes